Fix incorrect geocoding for addresses on streets
